+++ src/Linna/Session/PgsqlPdoSessionHandler.php
@@ -37,7 +37,8 @@ class PgsqlPdoSessionHandler implements SessionHandlerInterface
      * @param ExtendedPDO $pdo
      */
     public function __construct(private ExtendedPDO $pdo)
-    {}
+    {
+    }
 
     /**
      * Open session storage.